Move history pruning into NavigationHistory
